pyproject.toml
[tool.poetry]
name = "tp-authentication"
version = "1.0.0"
description = ""
authors = ["Behrooz Payoon <behrooz.payoon@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
